requirements.txt
anyio==3.6.1
ariadne==0.15.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
graphql-core==3.2.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
sniffio==1.2.0
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
